added debug window in settings and removed from setup
